refactor: move subcomponent styles to local scope files
