+++ src/views/components/GamePanel/GamePanel.module.scss
@@ -38,25 +38,8 @@
 					font-weight : 500;
 					display     : flex;
 					gap         : 1ch;
-
-					.Tags {
-						color       : palette.$black-d2;
-						display     : flex;
-						flex-wrap   : wrap;
-						font-weight : bold;
-						gap         : 1ch;
-						font-size   : 0.9rem;
-
-						.GenreTag {
-							background-color : palette.$white-d6;
-							border-radius    : 64px;
-							padding          : 0.375rem .75rem;
-							width            : fit-content;
-						}
-					}
 				}
 			}
-
 
 			.Description {
 				margin-block: 2rem;
