+++ src/views/components/GameTags/IGameTags.ts
